improve last.fm data display
