+++ client/actions/wikipediaActions.js
@@ -176,77 +176,3 @@ const containsArtistIdentifiers = (text) => {
 const doesNotContainDiscography = (text) => {
   return !containsStrings(text, [DISCOGRAPHY]);
 };
-
-const LAST_FM_ARTIST_GET_INFO =
-  'http://ws.audioscrobbler.com/2.0/?method=artist.getinfo';
-export const getArtistInfo = (artistName, artistIndex, timeRange) => {
-  return async (dispatch, getState) => {
-    function onSuccess(success) {
-      const payload = {
-        artistIndex,
-        artistName,
-        timeRange,
-        extract: success,
-      };
-      dispatch({ type: types.SPOTIFY_FETCH_ARTIST_SUCCESS, payload });
-      return success;
-    }
-    function onError(error) {
-      dispatch({ type: types.SPOTIFY_FETCH_ARTIST_FAILURE, error });
-      return error;
-    }
-
-    dispatch(loadArtistOverlay());
-
-    const state = getState();
-    const stateArtistName = state.selectedArtist.name;
-    // if artist is already loaded, then just open the overlay
-    if (stateArtistName === artistName) {
-      dispatch(openArtistOverlay());
-      return;
-    }
-
-    try {
-      const data = await fetchArtistInfo(artistName);
-      if (data) {
-        console.log(data);
-        let summary = '';
-        if (data && data.artist && data.artist.bio && data.artist.bio.summary) {
-          summary = data.artist.bio.summary;
-        }
-        return onSuccess(summary);
-      }
-    } catch (error) {
-      onError(error);
-    }
-
-    return onError('Artist info not found');
-  };
-};
-
-const fetchArtistInfo = async (artistName) => {
-  const pageEndpoint = `${LAST_FM_ARTIST_GET_INFO}&artist=${artistName}&api_key=${process.env.LAST_FM_API_KEY}&format=json`;
-  return await fetch(pageEndpoint)
-    .then((response) => response.json())
-    .then((data) => {
-      console.log('LAST FM');
-      console.log(data);
-      return data;
-    })
-    .catch((e) => console.log('An error occurred ' + e));
-};
-
-/*export const getArtist = async (artistName) => {
-  const pageEndpoint = `${LAST_FM_ARTIST_GET_INFO}&artist=${artistName}&api_key=${process.env.LAST_FM_API_KEY}&format=json`;
-  return await fetch(pageEndpoint)
-    .then((response) => response.json())
-    .then((data) => {
-      console.log('LAST FM');
-      console.log(data);
-      console.log(data.artist.tags.tag);
-      if (data && data.artist && data.artist.tags && data.artist.tags.tag) {
-        data.artist.tags.tag.map(item => item.name).join(", ")
-      }
-    })
-    .catch((e) => console.log('An error occurred ' + e));
-  };*/
