+++ package.json
@@ -50,6 +50,7 @@
     "halogenium": "^2.3.0",
     "history": "^1.17.0",
     "js-cookie": "^2.2.1",
+    "linkifyjs": "^2.1.9",
     "morgan": "^1.6.1",
     "node-vibrant": "^3.1.5",
     "object-assign": "^4.0.1",
